$auction delete ok

+++ src/Services/Database/AuctionsService.php
@@ -12,6 +12,7 @@
     use PluginAuctions\Constants\BidStatus;
     use PluginAuctions\Models\Auction_7;
     use PluginAuctions\Models\Fields\AuctionBidderListEntry;
+    use PluginAuctions\Models\VisitorCounter_1;
 
 //    use IO\Services\SessionStorageService;
 
@@ -301,28 +302,6 @@
             return false;
         }
 
-        /**
-         * @param $auctionId
-         * @return bool|mixed|string
-         */
-        public function getAuction($id)
-        {
-            if ($id > 0)
-            {
-                $auction = $this -> getValue(Auction_7::class, $id);
-
-                $this -> getLogger(__METHOD__)
-                      -> debug('PluginAuctions::Template.debug', ['$auction: ' => $auction]);
-
-                if ($auction instanceof Auction_7)
-                {
-                    return $this -> buildAuctionView($auction);
-                }
-            }
-
-            return 'falsche ID';
-        }
-
         public function getBidderList($id)
         {
             if ($id > 0)
@@ -573,8 +552,8 @@
 
                     $variationIds = $this -> itemService -> getVariationIds($auction -> itemId);
 
-                    $salesPriceData = ["variationId"  => $variationIds[0],
-                                       "price"        => $newEntry -> bidPrice,
+                    $salesPriceData = ["variationId" => $variationIds[0],
+                                       "price" => $newEntry -> bidPrice,
                                        "salesPriceId" => $salesPriceId
                     ];
 //
@@ -651,26 +630,52 @@
                 $auction = pluginApp(Auction_7::class);
                 $auction = $this -> getAuction($id);
 
-                $visitorCounter = $this -> visitorCounterService -> getVisitorCounterForItemId($auction -> itemId );
+                $visitorCounter = $this -> visitorCounterService -> getVisitorCounterForItemId($auction -> itemId);
 
                 $this -> getLogger(__METHOD__)
                       -> setReferenceType('testedId')
                       -> setReferenceValue($auction -> id)
                       -> debug('PluginAuctions::Template.debugBefor', ['auction: ' => $auction, 'visitorCounter' => $visitorCounter]);
 
-                $deleteVisitorCounter = $this -> visitorCounterService -> deleteVisitorCounter($visitorCounter -> id);
+                if ($visitorCounter instanceof VisitorCounter_1)
+                {
+                    $deleteVisitorCounter = $this -> visitorCounterService -> deleteVisitorCounter($visitorCounter -> id);
+                }
 
                 $this -> getLogger(__METHOD__)
                       -> setReferenceType('testedId')
                       -> setReferenceValue($visitorCounter -> id)
-                      -> debug('PluginAuctions::Template.debugAfter', ['deleteVisitorCounter: ' => $deleteVisitorCounter, 'auction'=> $auction]);
+                      -> debug('PluginAuctions::Template.debugAfter', ['deleteVisitorCounter: ' => $deleteVisitorCounter,
+                                                                       'auction'                => $auction
+                      ]);
 
 
-                return 'true';
-//                return json_encode($this -> deleteValue($auction));
+                return json_encode($this -> deleteValue($auction));
             }
 
             return 'Auctionsservice - delete Auction - Bedingung nicht erfüllt';
+        }
+
+        /**
+         * @param $auctionId
+         * @return bool|mixed|string
+         */
+        public function getAuction($id)
+        {
+            if ($id > 0)
+            {
+                $auction = $this -> getValue(Auction_7::class, $id);
+
+                $this -> getLogger(__METHOD__)
+                      -> debug('PluginAuctions::Template.debug', ['$auction: ' => $auction]);
+
+                if ($auction instanceof Auction_7)
+                {
+                    return $this -> buildAuctionView($auction);
+                }
+            }
+
+            return 'falsche ID';
         }
 
     }
